+++ packages/ui/src/HTML/React/style/HTMLElements/index.ts
@@ -1,3 +1,4 @@
 export * from "./Path";
 export * from "./P";
 export * from "./Input";
+export * from "./Img";
