--- packages/ui/src/HTML/React/style/Input.tsx
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export type InputStyle = {};
-
-export const Input = styled.input<InputStyle>`
-  width: 100%;
-  height: 100%;
-`;
