+++ packages/ui/src/HTML/React/style/index.ts
@@ -7,7 +7,6 @@ export * from "./Button";
 export * from "./Circle";
 export * from "./Div";
 export * from "./H1";
-export * from "./Img";
 
 export * from "./Areas/";
 export * from "./Loaders";
