+++ packages/ui/src/pages/styles/html-elements/img/index.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Img } from "../../../../HTML";
+
+export default function index() {
+  return <Img>LOLOLOL</Img>;
+}
